multiple tables creation

+++ bitbase_cli/src/executor/executor.cpp
@@ -1,32 +1,44 @@
 #include "executor/executor.h"
-#include "storage/table/table.h"
 #include <iostream>
-
-Executor::Executor(const char *filename)
-	: table(filename)
-{
-}
 
 void Executor::execute(const Statement &statement)
 {
 	switch (statement.type)
 	{
-	case (StatementType::INSERT):
+	case StatementType::INSERT:
 	{
+		Table *table = db.get_table(statement.table_name);
+
+		if (!table)
+		{
+			std::cout << "Error: Table not found\n";
+			break;
+		}
+
 		Row r;
 		r.id = statement.id;
 		r.username = statement.username;
 		r.email = statement.email;
 		r.age = statement.age;
 		r.is_active = statement.is_active;
-		table.insert(r);
-		std::cout << "Executed.\n";
+
+		table->insert(r);
+
+		std::cout << "Executed INSERT\n";
 		break;
 	}
 
-	case (StatementType::SELECT):
+	case StatementType::SELECT:
 	{
-		const auto &rows = table.get_all();
+		Table *table = db.get_table(statement.table_name);
+
+		if (!table)
+		{
+			std::cout << "Error: Table not found\n";
+			break;
+		}
+
+		const auto rows = table->get_all();
 
 		for (const auto &r : rows)
 		{
@@ -42,35 +54,86 @@ void Executor::execute(const Statement &statement)
 		break;
 	}
 
-	case (StatementType::DELETE):
+	case StatementType::DELETE:
 	{
-		bool deleted = table.delete_by_id(statement.id);
-		std::cout << "Executed.\n";
+		Table *table = db.get_table(statement.table_name);
+
+		if (!table)
+		{
+			std::cout << "Error: Table not found\n";
+			break;
+		}
+
+		bool deleted = table->delete_by_id(statement.id);
+
+		if (!deleted)
+		{
+			std::cout << "Row not found\n";
+		}
+		else
+		{
+			std::cout << "Executed DELETE\n";
+		}
+
 		break;
 	}
 
-	case (StatementType::UPDATE):
+	case StatementType::UPDATE:
 	{
+		Table *table = db.get_table(statement.table_name);
+
+		if (!table)
+		{
+			std::cout << "Error: Table not found\n";
+			break;
+		}
+
 		Row r;
 		r.id = statement.id;
 		r.username = statement.username;
 		r.email = statement.email;
 		r.age = statement.age;
 		r.is_active = statement.is_active;
-		table.update(r);
-		std::cout << "Executed.\n";
+
+		bool updated = table->update(r);
+
+		if (!updated)
+		{
+			std::cout << "Row not found\n";
+		}
+		else
+		{
+			std::cout << "Executed UPDATE\n";
+		}
+
 		break;
 	}
 
-	case (StatementType::CREATE_TABLE):
+	case StatementType::CREATE_TABLE:
 	{
-		std::cout << "Executed CREATE TABLE\n";
+		if (!db.create_table(statement.table_name))
+		{
+			std::cout << "Table already exists\n";
+		}
+		else
+		{
+			std::cout << "Executed CREATE TABLE\n";
+		}
+
 		break;
 	}
 
-	case (StatementType::DROP_TABLE):
+	case StatementType::DROP_TABLE:
 	{
-		std::cout << "Executed DROP TABLE\n";
+		if (!db.drop_table(statement.table_name))
+		{
+			std::cout << "Table not found\n";
+		}
+		else
+		{
+			std::cout << "Executed DROP TABLE\n";
+		}
+
 		break;
 	}
